Create README.md
About
Duplicate File Handler is a useful tool that can free some space on your drive. Write a handler that checks and compares files in a folder, displays the result, and removes duplicates.
Learning outcomes
Learn how to work with files and folders. Get familiar with hashing, learn how to apply it to your tasks.